Fix the broken unit tests
A couple of unit tests broke after the JSON representation was
simplified.
DB mappings are converted into a model by mimicking a decoder.
This broke when the JSON representation was simplified. The
decoding code has been updated to match the simplified JSON
representation. These changes along with fixes in the core
project fix the broken unit tests.

Resolves #7